fix bug in int, but float can be use also

+++ baseline/train.py
@@ -131,9 +131,10 @@ def softmax_model_pretrain(train_list, train_dir, class_count, target_model_path
     net.compile(optimizer=SGD(lr=0.001, momentum=0.9), loss='categorical_crossentropy', metrics=['accuracy'])
     net.fit_generator(
         train_datagen,
-        steps_per_epoch=train_cnt / 20 * 19 / batch_size + 1, epochs=40,
+        steps_per_epoch=int(train_cnt / 20 * 19 / batch_size + 1),
+        epochs=40,
         validation_data=val_datagen,
-        validation_steps=train_cnt / 20 / batch_size + 1,
+        validation_steps=int(train_cnt / 20 / batch_size + 1),
         callbacks=[save_best]
     )
     net.save(target_model_path)
